修改读表和生产系统UI问题以及增加建造系统与消耗材料相关的功能

+++ Assets/_ProjectOC/Scripts/Worker/Effect.cs
@@ -69,7 +69,7 @@ namespace ProjectOC.WorkerNS
                 #endregion
                 #region float
                 case EffectType.AlterWalkSpeed:
-                    ParamFloat = int.Parse(value);
+                    ParamFloat = float.Parse(value);
                     break;
                 #endregion
             }
